+++ examples/gas-station-demo/CHANGELOG.md
@@ -1,5 +1,15 @@
 # gas-station-demo
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [d70d947]
+  - @altude/core@0.1.1
+  - @altude/gasstation@2.1.0
+  - @altude/solana-adapter@0.0.1
+  - @altude/vault@0.0.1
+
 ## 0.0.5
 
 ### Patch Changes